Refactor

+++ ghsync/backup.py
@@ -1,5 +1,6 @@
 import click
 import shutil
+import os
 from datetime import datetime
 
 
@@ -7,6 +8,13 @@ def backup(dir):
     """
     Converts the directory into a zip file in the same parent directory with a timestamp and deletes the original directory.
     """
+    if not os.path.exists(dir):
+        click.echo(f"'{dir}' doesn't exist.")
+        return
+    if not os.path.isdir(dir):
+        click.echo(f"'{dir}' is not a directory.")
+        return
+
     timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
     shutil.make_archive(dir + "-" + timestamp, "zip", dir)
     shutil.rmtree(dir)
